DEV cfn
Create a Wowza server in EC2 with the documentation enabled (via launch config user data). Instance is locked down to Guardian office.

Why an EC2 instance and not a local Docker container or similar? Local instances are on a different internal subnet from Studio 1, so they cannot talk to each other.